backend/rag.py
from pydantic_ai import Agent

rag_agent = Agent(
    model='google-gla:gemini-2.5-flash',
    retries=1,
    system_prompt = (
        "You are a professional analyst. Answer the user's question strictly based on the provided context.",
        "If the answer is not found in the context, state that you do not know.",
        "Keep responses concise and direct. Do not use markdown formatting or code blocks unless requested."
    ),
)