pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai",
    "lancedb>=0.26.1",
    "pandas>=2.3.3",
    "pydantic-ai-slim>=1.44.0",
    "pypdf>=6.6.0",
    "streamlit>=1.53.0",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.12",
    "sentence-transformers>=3.0.0",
    "transformers>=4.30.0",
    "python-dotenv>=1.0.0",
    "passlib>=1.7.4",
    "pyjwt>=2.8.0",
    "requests>=2.32.0",
    "sqlalchemy>=2.0.0",
]
